Add colored circle test glyphs for v0 and v1 (#91)
* Add colored circle test glyphs for v0 and v1

Preparation for testing custom palette functionality. Introduce two
colored circle glyphs, one as COLRv0, one as COLRv1, going through the
colors of the rainbow from the outside to the inside.

Update logic inside test font generator script to allow adding COLRv0
glyphs to the COLR table, using COLRv0 style layer tuples as arguments
to buildCOLR().

Mark the new glyphs with an extra layer that shows a 0 or a 1 to
indicate the COLR glyph format (i.e. layers in v0, a graph of Paint
structures in v1).