Fix character mapping for extended accessor characters
Extended accessor characters were not correctly added to the iterator,
due to the use of the str() function instead of the chr()
function. So instead of greek letters, additional sequences of
numbers were added. This did not cause issues during testing so far
as the Skia tests were using glyph ids, not characters. But it is useful
to have character mappings for testing in the browser.

Update the font built, now exposing mappings to some greek characters
for the recently added glyphs.