PaintSweepGradient tests for coincident angles and stops
With a floating point implementation, tests for PaintVarSweep for
coincident angles and stops are difficult to achieve just by varying
variation parameters. It is more reliable to create separate static
glyphs for these.

This new set of tests adds a set of cases for testing coincident angles
and color stops, in forward and reverse order.