Add test glyphs for nested PaintGlyph
In order to test optimizations that recognize the sequence of
`PaintGlyph` and `Paint*Gradient`, create test glyphs that do not have
that exact sequence but instead express a nested clip by one
`PaintGlyph` being nested inside another. Add a transform that
transforms the inner `PaintGlyph` (and its gradient), and a transform
that transforms only the inner `Paint*Gradient`.